refactor: properly generate success response
